+++ test/faiss_test.rb
@@ -74,6 +74,27 @@ class FaissTest < Minitest::Test
     assert_equal [[0, 2, 1], [1, 2, 0], [2, 0, 1]], ids.to_a
   end
 
+  def test_index_hnsw_flat_inner_product
+    objects = [
+      [1, 1, 2, 1],
+      [5, 4, 6, 5],
+      [1, 2, 1, 2]
+    ]
+    index = Faiss::IndexHNSWFlat.new(4, 2, :inner_product)
+    index.add(objects)
+    distances, ids = index.search(objects, 3)
+
+    assert_equal [[26, 7, 7], [102, 29, 26], [29, 10, 7]], distances.to_a
+    assert_equal [[1, 2, 0], [1, 2, 0], [1, 2, 0]], ids.to_a
+  end
+
+  def test_invalid_metric
+    error = assert_raises do
+      Faiss::IndexHNSWFlat.new(4, 2, :bad)
+    end
+    assert_equal "Invalid metric: bad", error.message
+  end
+
   def test_index_ivf_flat
     objects = [
       [1, 1, 2, 1],